requirements.txt
Django==2.2.3
mysqlclient==1.4.2.post1
gunicorn==19.9.0
django-heroku==0.3.1
dj-database-url==0.5.0
whitenoise==4.1.3
django-mdeditor==0.1.14
markdown